Refactor panels and set auto-refresh of Environments panel
